requirements.txt
django-elasticsearch-dsl==0.4.0
django-environ==0.4.4
django-mptt==0.8.7
django-role-permissions==2.1.0
Django==1.11.5
djangorestframework==3.7.0
elasticsearch-dsl==5.3.0
elasticsearch==5.4.0
pdftotext==2.0.1
psycopg2==2.7.3.1
raven==6.2.1
pycountry==17.9.23
